@@ -1,7 +1,7 @@
 {
-  "name": "rs-ts-fastify",
+  "name": "nodejs-task-graphql-basics",
   "version": "1.0.0",
-  "description": "This project was bootstrapped with Fastify-CLI.",
+  "description": "nodejs-task-graphql-basics",
   "main": "app.ts",
   "nyc": {
     "sourceMap": false,
@@ -10,40 +10,70 @@
   "directories": {
     "test": "test"
   },
+  "type": "module",
+  "prisma": {
+    "seed": "ts-node --esm prisma/seed.ts"
+  },
   "scripts": {
-    "check-integrity": "node check-integrity.js",
-    "test": "npm run build:ts && tsc -p test/tsconfig.json && tap --ts \"test/**/*.test.ts\"",
+    "test-integrity": "node integrity.cjs",
+    "test-queries": "npm run build:ts && tsc -p test/tsconfig.json && tap --ts \"test/routes/gql-queries.test.ts\"",
+    "test-mutations": "npm run build:ts && tsc -p test/tsconfig.json && tap --ts \"test/routes/gql-mutations.test.ts\"",
+    "test-rule": "npm run build:ts && tsc -p test/tsconfig.json && tap --ts \"test/routes/gql-rule.test.ts\"",
+    "test-loader": "npm run build:ts && tsc -p test/tsconfig.json && tap --ts \"test/routes/gql-loader.test.ts\"",
+    "test-loader-prime": "npm run build:ts && tsc -p test/tsconfig.json && tap --ts \"test/routes/gql-loader-prime.test.ts\"",
     "start": "npm run build:ts && fastify start -l info dist/app.js",
     "build:ts": "tsc",
     "watch:ts": "tsc -w",
     "dev": "npm run build:ts && concurrently -k -p \"[{name}]\" -n \"TypeScript,App\" -c \"yellow.bold,cyan.bold\" \"npm:watch:ts\" \"npm:dev:start\"",
     "dev:start": "fastify start --ignore-watch=.ts$ -w -l info -P dist/app.js"
   },
+  "engines": {
+    "node": ">=18.16.0 <19.0.0"
+  },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@faker-js/faker": "^7.6.0",
     "@fastify/autoload": "^5.0.0",
+    "@fastify/env": "^4.2.0",
     "@fastify/sensible": "^5.0.0",
-    "@fastify/type-provider-json-schema-to-ts": "^2.1.1",
+    "@fastify/swagger": "^8.6.0",
+    "@fastify/swagger-ui": "^1.9.0",
+    "@fastify/type-provider-typebox": "^3.2.0",
+    "@prisma/client": "^4.16.0",
     "dataloader": "^2.1.0",
     "fastify": "^4.0.0",
     "fastify-cli": "^5.5.1",
     "fastify-plugin": "^4.0.0",
     "graphql": "^16.6.0",
     "graphql-depth-limit": "^1.1.0",
-    "lodash": "^4.17.21"
+    "graphql-parse-resolve-info": "^4.13.0"
   },
   "devDependencies": {
     "@types/graphql-depth-limit": "^1.1.3",
-    "@types/lodash": "^4.14.188",
     "@types/node": "^18.0.0",
     "@types/tap": "^15.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.59.11",
+    "@typescript-eslint/parser": "^5.59.11",
     "concurrently": "^7.0.0",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
     "fastify-tsconfig": "^1.0.1",
+    "npm": "^9.8.0",
+    "prettier": "^3.0.0",
+    "prisma": "^4.16.0",
     "tap": "^16.1.0",
     "ts-node": "^10.4.0",
-    "typescript": "^4.5.4"
+    "typescript": "^5.0.0"
+  },
+  "peerDependencies": {
+    "@sinclair/typebox": "^0.28.20"
+  },
+  "tap": {
+    "node-arg": [
+      "--no-warnings",
+      "--loader",
+      "ts-node/esm"
+    ]
   }
 }
