@@ -1,16 +1,32 @@
-import { join } from 'path';
-import AutoLoad from '@fastify/autoload';
+import { dirname, join } from 'node:path';
+import { fileURLToPath } from 'node:url';
+import AutoLoad, { AutoloadPluginOptions } from '@fastify/autoload';
 import { FastifyPluginAsync } from 'fastify';
 
-const app: FastifyPluginAsync = async (fastify): Promise<void> => {
-  fastify.register(AutoLoad, {
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const opts: Partial<AutoloadPluginOptions> = {
+  ignoreFilter: (path: string) => {
+    const isFileNested = (path.match(new RegExp('/', 'g')) ?? []).length > 1;
+    if (!isFileNested) {
+      return false;
+    }
+    return !path.endsWith('index.js');
+  },
+  forceESM: true,
+};
+
+const app: FastifyPluginAsync = async (fastify, _) => {
+  void fastify.register(AutoLoad, {
     dir: join(__dirname, 'plugins'),
-    options: {},
+    ...opts,
   });
 
-  fastify.register(AutoLoad, {
+  void fastify.register(AutoLoad, {
     dir: join(__dirname, 'routes'),
-    options: {},
+    routeParams: true,
+    ...opts,
   });
 };
 
