@@ -1,6 +1,4 @@
 import fp from 'fastify-plugin';
 import sensible from '@fastify/sensible';
 
-export default fp(async (fastify): Promise<void> => {
-  fastify.register(sensible);
-});
+export default fp(async (fastify) => fastify.register(sensible));
