@@ -1,81 +1,101 @@
-import { FastifyPluginAsyncJsonSchemaToTs } from '@fastify/type-provider-json-schema-to-ts';
-import { idParamSchema } from '../../utils/reusedSchemas';
-import { createPostBodySchema, changePostBodySchema } from './schema';
-import type { PostEntity } from '../../utils/DB/entities/DBPosts';
+import { FastifyPluginAsyncTypebox, Type } from '@fastify/type-provider-typebox';
+import {
+  changePostByIdSchema,
+  createPostSchema,
+  getPostByIdSchema,
+  postSchema,
+} from './schemas.js';
 
-const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
-  fastify
-): Promise<void> => {
-  fastify.get('/', async function (request, reply): Promise<PostEntity[]> {
-    return await fastify.db.posts.findMany();
+const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
+  const { prisma, httpErrors } = fastify;
+
+  fastify.route({
+    url: '/',
+    method: 'GET',
+    schema: {
+      response: {
+        200: Type.Array(postSchema),
+      },
+    },
+    async handler() {
+      return prisma.post.findMany();
+    },
   });
 
-  fastify.get(
-    '/:id',
-    {
-      schema: {
-        params: idParamSchema,
+  fastify.route({
+    url: '/:postId',
+    method: 'GET',
+    schema: {
+      ...getPostByIdSchema,
+      response: {
+        200: postSchema,
+        404: Type.Null(),
       },
     },
-    async function (request, reply): Promise<PostEntity> {
-      const post = await fastify.db.posts.findOne({
-        key: 'id',
-        equals: request.params.id,
+    async handler(req) {
+      const post = await prisma.post.findUnique({
+        where: {
+          id: req.params.postId,
+        },
       });
-      if (!post) throw fastify.httpErrors.createError(404, 'Post not found');
-
+      if (post === null) {
+        throw httpErrors.notFound();
+      }
       return post;
-    }
-  );
+    },
+  });
 
-  fastify.post(
-    '/',
-    {
-      schema: {
-        body: createPostBodySchema,
+  fastify.route({
+    url: '/',
+    method: 'POST',
+    schema: {
+      ...createPostSchema,
+      response: {
+        200: postSchema,
       },
     },
-    async function (request, reply): Promise<PostEntity> {
-      return await fastify.db.posts.create(request.body);
-    }
-  );
-
-  fastify.delete(
-    '/:id',
-    {
-      schema: {
-        params: idParamSchema,
-      },
-    },
-    async function (request, reply): Promise<PostEntity> {
-      const post = await fastify.db.posts.findOne({
-        key: 'id',
-        equals: request.params.id,
+    async handler(req) {
+      return prisma.post.create({
+        data: req.body,
       });
-      if (!post) throw fastify.httpErrors.createError(400, 'Post not found');
-      
-      return await fastify.db.posts.delete(request.params.id);
-    }
-  );
+    },
+  });
 
-  fastify.patch(
-    '/:id',
-    {
-      schema: {
-        body: changePostBodySchema,
-        params: idParamSchema,
+  fastify.route({
+    url: '/:postId',
+    method: 'PATCH',
+    schema: {
+      ...changePostByIdSchema,
+      response: {
+        200: postSchema,
       },
     },
-    async function (request, reply): Promise<PostEntity> {
-      const post = await fastify.db.posts.findOne({
-        key: 'id',
-        equals: request.params.id,
+    async handler(req) {
+      return prisma.post.update({
+        where: { id: req.params.postId },
+        data: req.body,
       });
-      if (!post) throw fastify.httpErrors.createError(400, 'Post not found');
+    },
+  });
 
-      return await fastify.db.posts.change(request.params.id, request.body);
-    }
-  );
+  fastify.route({
+    url: '/:postId',
+    method: 'DELETE',
+    schema: {
+      ...getPostByIdSchema,
+      response: {
+        204: Type.Void(),
+      },
+    },
+    async handler(req, reply) {
+      void reply.code(204);
+      await prisma.post.delete({
+        where: {
+          id: req.params.postId,
+        },
+      });
+    },
+  });
 };
 
 export default plugin;
